requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.2.3
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
urllib3==2.4.0
Werkzeug==2.2.2
supabase==2.15.0
python-dotenv==1.1.0
pytest==8.3.5
pytest-cov==6.1.1
coverage==7.8.0